safari-534.53-branch/LayoutTests/http/tests/security/frame-loading-via-document-write-expected.txt

CONSOLE MESSAGE: line 0: Not allowed to load local resource: abe.png


--------
Frame: 'topRow'
--------
This page was successfully loaded.
My protocol is http:
My referrer is blank


--------
Frame: 'middleRow'
--------
Image NOT loaded.


--------
Frame: 'bottomRow'
--------
Image loaded.
